refactor() Reviewing database modelling, services and configs: separate memory monitoring from conversion, provide metrics in audit models
